WIP: revert err msg for android
