+++ app/components/controls/picker-box-ios.js
@@ -14,6 +14,7 @@ export default class PickerBoxIos extends SafeComponent {
         super(props);
         this.focus = this.focus.bind(this);
         this.picker = this.props.picker;
+        this.opened = false;
     }
 
     focus() {
@@ -23,6 +24,21 @@ export default class PickerBoxIos extends SafeComponent {
             return;
         }
         uiState.showPicker(this.picker);
+    }
+
+
+    get errorSpacer() {
+        return (<View style={this.props.style.errorStyle} />);
+    }
+
+    get errorMessage() {
+        if (!this.props.value && this.opened) {
+            return (
+                <Text style={this.props.style.errorStyle}>
+                    {this.props.errorMessage}
+                </Text>);
+        }
+        return this.errorSpacer;
     }
 
     renderThrow() {
@@ -51,6 +67,7 @@ export default class PickerBoxIos extends SafeComponent {
                         </View>
                     </View>
                 </View>
+                {this.errorMessage}
             </View>
         );
     }
